Move TableExpander out of Services
