+++ src/Command/PutCommand.php
@@ -3,14 +3,12 @@
 namespace Meanbee\Magedbm2\Command;
 
 use Meanbee\Magedbm2\Application\ConfigInterface;
-use Meanbee\Magedbm2\Service\ConfigurableServiceInterface;
-use Meanbee\Magedbm2\Service\ConfigurationException;
+use Meanbee\Magedbm2\Helper\TableGroupExpander;
 use Meanbee\Magedbm2\Service\DatabaseInterface;
 use Meanbee\Magedbm2\Service\FilesystemInterface;
 use Meanbee\Magedbm2\Service\ServiceException;
 use Meanbee\Magedbm2\Service\StorageInterface;
 use Meanbee\Magedbm2\Service\TableExpanderInterface;
-use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
@@ -30,7 +28,7 @@ class PutCommand extends BaseCommand
     /** @var FilesystemInterface */
     protected $filesystem;
 
-    /** @var TableExpanderInterface */
+    /** @var TableGroupExpander */
     protected $tableExpander;
     
     /** @var ConfigInterface */
@@ -41,14 +39,14 @@ class PutCommand extends BaseCommand
      * @param DatabaseInterface $database
      * @param StorageInterface $storage
      * @param FilesystemInterface $filesystem
-     * @param TableExpanderInterface $tableExpander
+     * @param TableGroupExpander $tableGroupExpander
      */
     public function __construct(
         ConfigInterface $config,
         DatabaseInterface $database,
         StorageInterface $storage,
         FilesystemInterface $filesystem,
-        TableExpanderInterface $tableExpander
+        TableGroupExpander $tableGroupExpander = null
     )
     {
         parent::__construct();
@@ -56,7 +54,7 @@ class PutCommand extends BaseCommand
         $this->database = $database;
         $this->storage = $storage;
         $this->filesystem = $filesystem;
-        $this->tableExpander = $tableExpander;
+        $this->tableExpander = $tableGroupExpander ?? new TableGroupExpander();
         $this->config = $config;
 
         $this->ensureServiceConfigurationValidated('database', $this->database);
